jshint linter: add option to ignore specific error messages
*warning: very gross approach*

This commit adds the "lint.jshint.errorstoignore" option to arcconfig,
as an array, so that a user is now able to ignore any specific error
messages he wishes, by listing it there.

i.e. (in .arcconfig):

+---------------------------------------------------------------------+
| "lint.jshint.errorstoignore": [ "Missing space after 'function'." ] |
+---------------------------------------------------------------------+

will ignore the annoying "Missing space after 'function'" warning, which
complains about constructions in the form of "foo = function() {}".

I am not aware of a way to disable that checking, without also disabling
the "indent" option, which seems to be often undesirable.

Likely, the proper way of fixing that specific thing is to be implemented
in jshint itself, but it seemed interesting to have a way to supress an
arbitrary error/warning in arcanist.

Signed-off-by: Sergio Correia <sergio@correia.cc>